requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
discord.py==2.3.2
frozenlist==1.4.1
gTTS==2.5.1
idna==3.7
multidict==6.0.5
requests==2.32.3
soupsieve==2.5
urllib3==2.2.1
yarl==1.9.4
gunicorn
PyNaCl